requirements.txt
streamlit
requests
beautifulsoup4
python-dotenv
langchain
langchain-openai
llama-index
